split from vm-shim.js to create find-global-declarations.js
