+++ lib/jsdom/vm-shim.js
@@ -4,7 +4,7 @@ const acorn = require("acorn");
 const findGlobals = require("acorn-globals");
 const escodegen = require("escodegen");
 const jsGlobals = require("./browser/js-globals.json");
-const walk = require("acorn-walk");
+const { findGlobalDeclarations } = require("./find-global-declarations.js");
 
 // We can't use the default browserify vm shim because it doesn't work in a web worker.
 
@@ -16,73 +16,6 @@ const jsGlobalEntriesToInstall = Object.entries(jsGlobals).filter(([name]) => na
 // The implemention of this vm-shim.js is to enclose the code into function scope,
 // so all global declarations become local declarations, and then these new installation are dismissed.
 // The patch is assignment being used to install these variables into global object.
-
-function isFunctionScope(node) {
-  return node.type === "FunctionExpression" || node.type === "FunctionDeclaration" ||
-    node.type === "ArrowFunctionExpression";
-}
-
-function findGlobalDeclarations(ast) {
-  const globalDeclarations = [];
-  function declarePattern(node, varNode) {
-    switch (node.type) {
-      case "Identifier":
-        varNode.globals[node.name] = true;
-        break;
-      case "ObjectPattern":
-        node.properties.forEach(property => {
-          declarePattern(property.value || property.argument, varNode);
-        });
-        break;
-      case "ArrayPattern":
-        node.elements.forEach(element => {
-          if (element) {
-            declarePattern(element, varNode);
-          }
-        });
-        break;
-      case "RestElement":
-        declarePattern(node.argument, varNode);
-        break;
-      case "AssignmentPattern":
-        declarePattern(node.left, varNode);
-        break;
-      default:
-        throw new Error("Unrecognized pattern type: " + node.type);
-    }
-  }
-  walk.ancestor(ast, {
-    VariableDeclaration(node, parents) {
-      if (node.kind !== "var") {
-        return;
-      }
-      for (let i = parents.length - 2; i >= 0; i--) {
-        if (isFunctionScope(parents[i])) {
-          return;
-        }
-      }
-      node.globals = node.globals || Object.create(null);
-      node.declarations.forEach(declaration => {
-        declarePattern(declaration.id, node);
-      });
-      globalDeclarations.push(node);
-    },
-    FunctionDeclaration(node, parents) {
-      for (let i = parents.length - 2; i >= 0; i--) {
-        if (isFunctionScope(parents[i])) {
-          return;
-        }
-      }
-      node.globals = node.globals || Object.create(null);
-      if (node.id) {
-        node.globals[node.id.name] = true;
-      }
-      node.parent = parents[parents.length - 2];
-      globalDeclarations.push(node);
-    }
-  });
-  return globalDeclarations;
-}
 
 function getPatchNode(name) {
   return {
